+++ changelog.md
@@ -1,5 +1,6 @@
 - Initial port to 1.21.1
 - contains saccharine tree and apricorn trees
-  Missing features:
-- slathering saccharine tree with honey
-- slathering saccharine leaves with honey
+- Saccharine trees are fully functional (honey slathering, honey leaves, etc).
+
+Missing feature:
+- Saccharine trees normally have a chance to spawn with combees instead of bees. This has not been implemented yet.